\name{print.treeClust}
\alias{print.treeClust}
\title{
Print treeClust object
}
\description{
Print some details about a treeClust object, and the "tbl" element.
}
\usage{
\method{print}{treeClust} (x, \ldots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Object of class treeClust
}
  \item{\dots}{
Other stuff.
}
}
\value{
None. The "tbl" element is printed to the screen.
}
