\name{summary.treeClust}
\alias{summary.treeClust}
\title{
Summarize treeClust object
}
\description{
Print some details about a treeClust object.
}
\usage{
\method{summary}{treeClust} (object, \ldots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
Object of class treeClust
}
  \item{\dots}{
Other stuff.
}
}
\value{
None. A few lines of information are printed to the screen.
}
